# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is backend api service for tks platform",
        "title": "tks-api service",
        "contact": {
            "name": "taekyu.kang@sk.com",
            "email": "taekyu.kang@sk.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "tks-api-dev.taco-cat.xyz:9110",
    "basePath": "/api/1.0/",
    "paths": {
...
        "/organizations/{organizationId}/users": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get user list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organizationId",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.ListUserBody"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organizationId",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "create user request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "create user response",
                        "schema": {
                            "$ref": "#/definitions/domain.CreateUserResponse"
                        }
                    }
                }
            }
        },
        "/organizations/{organizationId}/users/{accountId}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get user detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organizationId",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "accountId",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetUserResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update user detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organizationId",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "accountId",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update user request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateUserResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Delete user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organizationId",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "accountId",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    }
                }
            }
        },
        "/organizations/{organizationId}/users/{accountId}/existence": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "return true when accountId exists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user id existence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organizationId",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "accountId",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/organizations/{organizationId}/users/{accountId}/password": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update user password detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user password detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organizationId",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "accountId",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update user password request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UpdatePasswordResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateUserRequest": {
            "type": "object",
            "required": [
                "accountId",
                "password",
                "role"
            ],
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "department": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 0
                },
                "description": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 0
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "user"
                    ]
                }
            }
        },
        "domain.CreateUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "accountId": {
                            "type": "string"
                        },
                        "department": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "id": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "organization": {
                            "$ref": "#/definitions/domain.Organization"
                        },
                        "role": {
                            "$ref": "#/definitions/domain.Role"
                        }
                    }
                }
            }
        },
...
        "domain.GetUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "accountId": {
                            "type": "string"
                        },
                        "createdAt": {
                            "type": "string"
                        },
                        "creator": {
                            "type": "string"
                        },
                        "department": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "id": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "organization": {
                            "$ref": "#/definitions/domain.Organization"
                        },
                        "role": {
                            "$ref": "#/definitions/domain.Role"
                        },
                        "updatedAt": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "domain.ListUserBody": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/domain.Organization"
                },
                "role": {
                    "$ref": "#/definitions/domain.Role"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "domain.Organization": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "primaryClusterId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.OrganizationStatus"
                },
                "statusDescription": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.OrganizationStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7
            ],
            "x-enum-varnames": [
                "OrganizationStatus_PENDING",
                "OrganizationStatus_CREATE",
                "OrganizationStatus_CREATING",
                "OrganizationStatus_CREATED",
                "OrganizationStatus_DELETE",
                "OrganizationStatus_DELETING",
                "OrganizationStatus_DELETED",
                "OrganizationStatus_ERROR"
            ]
        },
        "domain.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "domain.UpdatePasswordResponse": {
            "type": "object"
        },
...
        "domain.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 0
                },
                "description": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 0
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 0
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "user"
                    ]
                }
            }
        },
        "domain.UpdateUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "accountId": {
                            "type": "string"
                        },
                        "department": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "id": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "organization": {
                            "$ref": "#/definitions/domain.Organization"
                        },
                        "role": {
                            "$ref": "#/definitions/domain.Role"
                        }
                    }
                }
            }
        },
        "domain.User": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/domain.Organization"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/domain.Role"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}